requirements.txt
pandas
numpy
MetaTrader5
ta
pyyaml
